requirements.txt
tensorflow
streamlit
pillow
matplotlib
opencv-python
